+++ backend/src/baserow/contrib/database/api/fields/views.py
@@ -21,11 +21,13 @@ from baserow.contrib.database.api.fields.errors import (
     ERROR_CANNOT_DELETE_PRIMARY_FIELD,
     ERROR_CANNOT_CHANGE_FIELD_TYPE,
     ERROR_FIELD_DOES_NOT_EXIST,
+    ERROR_MAX_FIELD_COUNT_EXCEEDED,
 )
 from baserow.contrib.database.fields.exceptions import (
     CannotDeletePrimaryField,
     CannotChangeFieldType,
     FieldDoesNotExist,
+    MaxFieldLimitExceeded,
 )
 from baserow.contrib.database.fields.models import Field
 from baserow.contrib.database.fields.handler import FieldHandler
@@ -122,7 +124,11 @@ class FieldsView(APIView):
                 field_type_registry, FieldSerializer
             ),
             400: get_error_schema(
-                ["ERROR_USER_NOT_IN_GROUP", "ERROR_REQUEST_BODY_VALIDATION"]
+                [
+                    "ERROR_USER_NOT_IN_GROUP",
+                    "ERROR_REQUEST_BODY_VALIDATION",
+                    "ERROR_MAX_FIELD_COUNT_EXCEEDED",
+                ]
             ),
             404: get_error_schema(["ERROR_TABLE_DOES_NOT_EXIST"]),
         },
@@ -135,6 +141,7 @@ class FieldsView(APIView):
         {
             TableDoesNotExist: ERROR_TABLE_DOES_NOT_EXIST,
             UserNotInGroup: ERROR_USER_NOT_IN_GROUP,
+            MaxFieldLimitExceeded: ERROR_MAX_FIELD_COUNT_EXCEEDED,
         }
     )
     def post(self, request, data, table_id):