+++ web-frontend/modules/core/plugins/clientHandler.js
@@ -56,6 +56,10 @@ class ErrorHandler {
         'Invalid URL',
         'The provided file URL is invalid or not allowed.'
       ),
+      ERROR_MAX_FIELD_COUNT_EXCEEDED: new ResponseErrorMessage(
+        "Couldn't create field.",
+        "The action couldn't be completed because the field count exceeds the limit"
+      ),
     }
 
     // A temporary notFoundMap containing the error messages for when the