+++ changelog.md
@@ -1,7 +1,9 @@
+
 # Changelog
 
 ## Unreleased
 
+* Added configurable field limit.
 * Fixed memory leak in the `link_row` field.
 * Switch to using a celery based email backend by default.
 * Added `--add-columns` flag to the `fill_table` management command. It creates all the